+++ components/camel-file-watch/src/test/java/org/apache/camel/component/file/watch/FileWatchComponentDeleteTest.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.camel.component.file.watch;
+
+import java.nio.file.Files;
+
+import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.component.mock.MockEndpoint;
+import org.junit.jupiter.api.Test;
+
+public class FileWatchComponentDeleteTest extends FileWatchComponentTestBase {
+
+    @Test
+    public void testRemoveFile() throws Exception {
+        MockEndpoint mock = getMockEndpoint("mock:watchDelete");
+        mock.expectedMessageCount(2);
+
+        Files.delete(testFiles.get(0));
+        Files.delete(testFiles.get(1));
+
+        mock.assertIsSatisfied();
+    }
+
+    @Override
+    protected RouteBuilder createRouteBuilder() throws Exception {
+        return new RouteBuilder() {
+            public void configure() {
+                from("file-watch://" + testPath() + "?events=DELETE")
+                        .to("mock:watchDelete");
+            }
+        };
+    }
+}
